Update javascript and expose release data

+++ Battlelogium.Core/Utilities/Process.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.Management;
 using System.Runtime.InteropServices;
 using System.Drawing;
+using System.Threading.Tasks;
 
 
 namespace Battlelogium.Core.Utilities
@@ -46,16 +47,14 @@ namespace Battlelogium.Core.Utilities
         {
             try
             {
-                if (closeMainWindow)
+                switch (closeMainWindow)
                 {
-                    process.CloseMainWindow();
-                }
-                else
-                {
-                    if (!process.HasExited)
-                    {
-                        process.Kill();
-                    }
+                    case true:
+                        process.CloseMainWindow();
+                        break;
+                    case false:
+                        if (!process.HasExited) process.Kill();
+                        break;
                 }
             }
             catch (Exception)
@@ -76,18 +75,21 @@ namespace Battlelogium.Core.Utilities
             }
         }
 
-
+        public static bool IsProcessRunning(string processName)
+        {
+            Process[] proc = Process.GetProcessesByName(processName);
+            if (proc.Length.Equals(0)) return false;
+            else return true;
+        }
     }
 
     public class ProcessStartWaiter
     {
         public string processName;
 
-        private bool loop = true;
         public delegate void ProcessStartEventHandler(object sender, ProcessStartEventArgs e);
         public event ProcessStartEventHandler ProcessStart;
        
-        // Adapted from http://stackoverflow.com/questions/6575117/
         public ProcessStartWaiter(string processName)
         {
             this.processName = processName;
@@ -95,40 +97,26 @@ namespace Battlelogium.Core.Utilities
 
         public void Listen()
         {
-            var query = new WqlEventQuery(
-                "__InstanceCreationEvent",
-                new TimeSpan(0, 0, 1),
-                "TargetInstance isa \"Win32_Process\" and TargetInstance.Name = '" + processName + "'"
-              );
+            string queryString =
+            "SELECT TargetInstance" +
+            "  FROM __InstanceCreationEvent " +
+            "WITHIN  10 " +
+            " WHERE TargetInstance ISA 'Win32_Process' " +
+            "   AND TargetInstance.Name = '" + processName + "'";
 
-            using (var watcher = new ManagementEventWatcher(query)){
-                while (this.loop)
-                {
-                    ManagementBaseObject wmiEvent = watcher.WaitForNextEvent();
-                    ManagementBaseObject targetInstance = (ManagementBaseObject)wmiEvent["targetInstance"];
-                    int pid = Convert.ToInt32(targetInstance["processID"]);
-                    Process process = Process.GetProcessById(pid);
-                    if(this.loop) this.OnProcessStart(new ProcessStartEventArgs(process, this.processName)); //Don't trigger event if we want to stop sending them.
-                    wmiEvent.Dispose();
-                    targetInstance.Dispose();
-                    process.Dispose();
-                }
-                watcher.Stop();
-            }
-        }
-
-        public void StopListen()
-        {
-            this.loop = false;
+            ManagementEventWatcher watcher = new ManagementEventWatcher(queryString);
+            watcher.EventArrived += (s, e) =>
+            {
+                ManagementBaseObject targetInstance = (ManagementBaseObject) e.NewEvent["targetInstance"];
+                int pid = Convert.ToInt32(targetInstance["processID"]);
+                OnProcessStart(new ProcessStartEventArgs(pid, this.processName));
+            };
+            watcher.Start();
         }
 
         public void ListenAsync()
         {
-            using (var worker = new BackgroundWorker())
-            {
-                worker.DoWork += delegate { this.Listen(); };
-                worker.RunWorkerAsync();
-            }
+            Task.Run(() => this.Listen());
         }
 
         private void OnProcessStart(ProcessStartEventArgs e)
@@ -140,14 +128,13 @@ namespace Battlelogium.Core.Utilities
 
     public class ProcessStartEventArgs : EventArgs
     {
-        public Process Process { get; private set; }
+        public int ProcessID { get; private set; }
         public string ProcessName { get; private set; }
 
-        public ProcessStartEventArgs(Process process, string processName)
+        public ProcessStartEventArgs(int processID, string processName)
         {
-            Process = process;
+            ProcessID = processID;
             ProcessName = processName;
-            
         }
     }
